legacy/spark-for-data-enginners-main/pyproject.toml
[tool.poetry]
name = "spark-for-data-engginers"
version = "0.1.0"
description = ""
authors = ["Guilherme S. Magalhães <40049979+Guisilcol@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "spark_for_data_engginers"}]

[tool.poetry.dependencies]
python = "^3.10"
names = "^0.3.0"
kafka-python = "^2.0.2"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.19.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
